agentic-file-recommender/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sentence-transformers>=2.2.2
torch>=2.2.0
tensorflow>=2.15.0
transformers>=4.30.2
annoy>=1.17.0
pyyaml>=5.4.1
python-multipart>=0.0.5
chardet>=4.0.0
